requirements.txt
flask
yfinance
pandas
ta
feedparser
transformers
torch
pytz
python-dateutil
